+++ src/Covid19Api.Presentation/Response/VaccinationStatisticValueDto.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace Covid19Api.Presentation.Response
 {
-    public record VaccinationStatisticValueDto(DateTime LoggedAt, uint TotalVaccinations, uint PeopleVaccinated, uint PeopleFullyVaccinated, double PeopleFullyVaccinatedPerHundred, uint DailyVaccinations, double TotalVaccinationsPerHundred, double PeopleVaccinatedPerHundred, uint DailyVaccinationsPerMillion)
+    public record VaccinationStatisticValueDto(DateTime LoggedAt, ulong TotalVaccinations, ulong PeopleVaccinated, ulong PeopleFullyVaccinated, double PeopleFullyVaccinatedPerHundred, ulong DailyVaccinations, double TotalVaccinationsPerHundred, double PeopleVaccinatedPerHundred, uint DailyVaccinationsPerMillion)
     {
     }
 }
